obrasAPI/requirements.txt
Flask==1.1.4
requests==2.23.0
responses==0.10.14
python-dotenv==0.17.1